refactor!: Move components to the `optimum` submodule
