+++ integrations/optimum/src/haystack_integrations/components/embedders/optimum/__init__.py
