+++ integrations/optimum/src/haystack_integrations/components/embedders/optimum/optimum_backend.py
@@ -3,7 +3,7 @@ from typing import Any, Dict, List, Optional, Union
 import numpy as np
 import torch
 from haystack.utils.auth import Secret
-from haystack_integrations.components.embedders.pooling import Pooling, PoolingMode
+from .pooling import Pooling, PoolingMode
 from optimum.onnxruntime import ORTModelForFeatureExtraction
 from tqdm import tqdm
 from transformers import AutoTokenizer
