+++ integrations/optimum/src/haystack_integrations/components/embedders/optimum/optimum_text_embedder.py
@@ -1,34 +1,30 @@
 from typing import Any, Dict, List, Optional, Union
 
-from haystack import Document, component, default_from_dict, default_to_dict
+from haystack import component, default_from_dict, default_to_dict
 from haystack.utils import Secret, deserialize_secrets_inplace
 from haystack.utils.hf import HFModelType, check_valid_model, deserialize_hf_model_kwargs, serialize_hf_model_kwargs
-from haystack_integrations.components.embedders.optimum_backend import OptimumEmbeddingBackend
-from haystack_integrations.components.embedders.pooling import HFPoolingMode, PoolingMode
+from .optimum_backend import OptimumEmbeddingBackend
+from .pooling import HFPoolingMode, PoolingMode
 
 
 @component
-class OptimumDocumentEmbedder:
+class OptimumTextEmbedder:
     """
-    A component for computing Document embeddings using models loaded with the HuggingFace Optimum library.
+    A component to embed text using models loaded with the HuggingFace Optimum library.
     This component is designed to seamlessly inference models using the high speed ONNX runtime.
-
-    The embedding of each Document is stored in the `embedding` field of the Document.
 
     Usage example:
     ```python
-    from haystack.dataclasses import Document
-    from haystack_integrations.components.embedders import OptimumDocumentEmbedder
+    from haystack_integrations.components.optimum.embedders import OptimumTextEmbedder
 
-    doc = Document(content="I love pizza!")
+    text_to_embed = "I love pizza!"
 
-    document_embedder = OptimumDocumentEmbedder(model="sentence-transformers/all-mpnet-base-v2")
-    document_embedder.warm_up()
+    text_embedder = OptimumTextEmbedder(model="sentence-transformers/all-mpnet-base-v2")
+    text_embedder.warm_up()
 
-    result = document_embedder.run([doc])
-    print(result["documents"][0].embedding)
+    print(text_embedder.run(text_to_embed))
 
-    # [0.017020374536514282, -0.023255806416273117, ...]
+    # {'embedding': [-0.07804739475250244, 0.1498992145061493,, ...]}
     ```
 
     Key Features and Compatibility:
@@ -54,13 +50,9 @@ class OptimumDocumentEmbedder:
         onnx_execution_provider: str = "CPUExecutionProvider",
         pooling_mode: Optional[Union[str, PoolingMode]] = None,
         model_kwargs: Optional[Dict[str, Any]] = None,
-        batch_size: int = 32,
-        progress_bar: bool = True,
-        meta_fields_to_embed: Optional[List[str]] = None,
-        embedding_separator: str = "\n",
     ):
         """
-        Create a OptimumDocumentEmbedder component.
+        Create a OptimumTextEmbedder component.
 
         :param model: A string representing the model id on HF Hub.
         :param token: The HuggingFace token to use as HTTP bearer authorization.
@@ -77,7 +69,7 @@ class OptimumDocumentEmbedder:
             recommend setting these two provider options using the model_kwargs parameter, when using the TensorRT
             execution provider. The usage is as follows:
             ```python
-            embedder = OptimumDocumentEmbedder(
+            embedder = OptimumTextEmbedder(
                 model="sentence-transformers/all-mpnet-base-v2",
                 onnx_execution_provider="TensorrtExecutionProvider",
                 model_kwargs={
@@ -104,11 +96,6 @@ class OptimumDocumentEmbedder:
         :param model_kwargs: Dictionary containing additional keyword arguments to pass to the model.
             In case of duplication, these kwargs override `model`, `onnx_execution_provider`, and `token` initialization
             parameters.
-        :param batch_size: Number of Documents to encode at once.
-        :param progress_bar: Whether to show a progress bar or not. Can be helpful to disable in production deployments
-            to keep the logs clean.
-        :param meta_fields_to_embed: List of meta fields that should be embedded along with the Document text.
-        :param embedding_separator: Separator used to concatenate the meta fields to the Document text.
         """
         check_valid_model(model, HFModelType.EMBEDDING, token)
         self.model = model
@@ -134,10 +121,6 @@ class OptimumDocumentEmbedder:
         self.suffix = suffix
         self.normalize_embeddings = normalize_embeddings
         self.onnx_execution_provider = onnx_execution_provider
-        self.batch_size = batch_size
-        self.progress_bar = progress_bar
-        self.meta_fields_to_embed = meta_fields_to_embed or []
-        self.embedding_separator = embedding_separator
 
         model_kwargs = model_kwargs or {}
 
@@ -170,10 +153,6 @@ class OptimumDocumentEmbedder:
             normalize_embeddings=self.normalize_embeddings,
             onnx_execution_provider=self.onnx_execution_provider,
             pooling_mode=self.pooling_mode.value,
-            batch_size=self.batch_size,
-            progress_bar=self.progress_bar,
-            meta_fields_to_embed=self.meta_fields_to_embed,
-            embedding_separator=self.embedding_separator,
             model_kwargs=self.model_kwargs,
             token=self.token.to_dict() if self.token else None,
         )
@@ -185,7 +164,7 @@ class OptimumDocumentEmbedder:
         return serialization_dict
 
     @classmethod
-    def from_dict(cls, data: Dict[str, Any]) -> "OptimumDocumentEmbedder":
+    def from_dict(cls, data: Dict[str, Any]) -> "OptimumTextEmbedder":
         """
         Deserialize this component from a dictionary.
         """
@@ -194,36 +173,18 @@ class OptimumDocumentEmbedder:
         deserialize_hf_model_kwargs(data["init_parameters"]["model_kwargs"])
         return default_from_dict(cls, data)
 
-    def _prepare_texts_to_embed(self, documents: List[Document]) -> List[str]:
+    @component.output_types(embedding=List[float])
+    def run(self, text: str):
         """
-        Prepare the texts to embed by concatenating the Document text with the metadata fields to embed.
+        Embed a string.
+
+        :param text: The text to embed.
+        :return: The embeddings of the text.
         """
-        texts_to_embed = []
-        for doc in documents:
-            meta_values_to_embed = [
-                str(doc.meta[key]) for key in self.meta_fields_to_embed if key in doc.meta and doc.meta[key] is not None
-            ]
-
-            text_to_embed = (
-                self.prefix + self.embedding_separator.join([*meta_values_to_embed, doc.content or ""]) + self.suffix
-            )
-
-            texts_to_embed.append(text_to_embed)
-        return texts_to_embed
-
-    @component.output_types(documents=List[Document])
-    def run(self, documents: List[Document]):
-        """
-        Embed a list of Documents.
-        The embedding of each Document is stored in the `embedding` field of the Document.
-
-        :param documents: A list of Documents to embed.
-        :return: A dictionary containing the updated Documents with their embeddings.
-        """
-        if not isinstance(documents, list) or documents and not isinstance(documents[0], Document):
+        if not isinstance(text, str):
             msg = (
-                "OptimumDocumentEmbedder expects a list of Documents as input."
-                " In case you want to embed a string, please use the OptimumTextEmbedder."
+                "OptimumTextEmbedder expects a string as an input. "
+                "In case you want to embed a list of Documents, please use the OptimumDocumentEmbedder."
             )
             raise TypeError(msg)
 
@@ -231,21 +192,10 @@ class OptimumDocumentEmbedder:
             msg = "The embedding model has not been loaded. Please call warm_up() before running."
             raise RuntimeError(msg)
 
-        # Return empty list if no documents
-        if not documents:
-            return {"documents": []}
+        text_to_embed = self.prefix + text + self.suffix
 
-        texts_to_embed = self._prepare_texts_to_embed(documents=documents)
-
-        embeddings = self.embedding_backend.embed(
-            texts_to_embed=texts_to_embed,
-            normalize_embeddings=self.normalize_embeddings,
-            pooling_mode=self.pooling_mode,
-            progress_bar=self.progress_bar,
-            batch_size=self.batch_size,
+        embedding = self.embedding_backend.embed(
+            texts_to_embed=text_to_embed, normalize_embeddings=self.normalize_embeddings, pooling_mode=self.pooling_mode
         )
 
-        for doc, emb in zip(documents, embeddings):
-            doc.embedding = emb
-
-        return {"documents": documents}
+        return {"embedding": embedding}
